Replace cryptsetup with cryptmount wrapper
Cryptmount provides a lot of the work that is duplicate in the
mkinitcpio encrypt hook and initscripts. Use it to map all volumes from
crypttab that don't have "noauto" and which aren't already mounted.
Also, use it to unmap volumes in crypttab that aren't already mounted
and aren't / or /usr.