Skip cryptos tagged as %early
We'll have the late hook in the initrd automatically map all volumes
with an %early tag. Skipping them here is up for discussion though, as
they should be skipped if they're mapped anyway.

What should the procedure be if we're in rc.sysinit, and an %early
volume is not mounted for whatever reason?