rc.sysinit: remove warnings
These were useful in a transition period, but no need to keep them around.

If /usr is not mounted we probably won't even get this far. If devtmpfs is not supported
the previous mount of /dev would have given an errormessage.

This was prompted by a suggeston from Igor Podlesny to improve on the error handling, which
made me realise we could drop it.

Signed-off-by: Tom Gundersen <teg@jklm.no>