Correct spelling and grammar throughout
This patch addresses the issues with my previous patch submission here:
https://mailman.archlinux.org/pipermail/arch-projects/2012-June/002990.html

The notable changes between this patch and the previous one are below:
-- this patch does not rename set_timezone()
-- 'filesystems' and 'bash' are left as-is
-- everything is in one patch, not two
-- the 'makedepends' change is no longer relevant because the duplication
of it was already removed in a later commit

Regarding the capitalization of "Btrfs", it seems that upstream prefers
"Btrfs".[1][2]

Regarding 'https://' vs. 'http://', I left this change in the patch
because stating 'https://www.archlinux.org' in the start-up messages
doesn't force anyone to use HTTPS over HTTP. It simply shows that HTTPS
is available for use. If someone really doesn't want to use HTTPS, they
don't have to. Finally, Dan McGee stated that HTTPS will be used all the
time soon.[3]

[1] https://oss.oracle.com/projects/btrfs/
[2] https://btrfs.wiki.kernel.org/index.php/Main_Page
[3] https://bugs.archlinux.org/task/30486#comment96222

Signed-off-by: Jason St. John <jstjohn@purdue.edu>