rc.conf: prune
The valid options are documented in rc.conf(5). Keeping everything in the manpage means we don't
have to deal with .pacnew files when updating minor detals (such as default options).

Furthermore, all the removed options are recommended to be left unset. Only if you have special
needs should you need to change them.

This rc.conf file can be used unchanged on a basic system (with network, logging and cron).

Signed-off-by: Tom Gundersen <teg@jklm.no>